fix: 구조 분리 및 탭 스크롤 수정
